media/requirements.txt
# FastAPI Media Service Dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
motor==3.6.0
pymongo==4.9.1
boto3==1.35.0
pillow
python-magic
redis
pydantic
pydantic-settings
python-jose[cryptography]==3.3.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

